refactor: 폴더 구조 개편에 따른 파일 이동
